backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
geopandas>=0.14.2
matplotlib>=3.8.0
python-pptx>=0.6.23
pandas>=2.1.0
contextily>=1.5.0
pillow>=10.0.0
shapely>=2.0.0
geopy>=2.4.0
pyproj>=3.6.0
